Service.Device/requirements.txt
flask
flask-pymongo
pydantic
fastapi
flask-cors
requests
gevent
pymongo
python-icc
